apps/sinornis/requirements.txt
# Core dependencies for poultry alert system
flask>=2.3.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Sensor and hardware communication
pyserial>=3.5
RPi.GPIO>=0.7.1; platform_machine=="armv7l"
gpiozero>=1.6.2; platform_machine=="armv7l"

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Database
sqlite3

# Messaging and notifications
smtplib
twilio>=8.0.0

# Scheduling and automation
schedule>=1.2.0
APScheduler>=3.10.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0